Always return mimetype as string
Currently MultiMime.lookup returns either 
::MIME::Type if it uses MIME::Types or ::Mime::Type
and returns a string containing the content type
if it uses ::Rack::Mime. This change makes the 
return type consistently a string, which is 
supported by all fog-providers. This change also 
allows us to remove the unnecessary type checking 
in storage.rb as we can be assured its a string.