Adds JSON API support when using Google Storage
The current Google Storage support requires the
use of the S3 compatible API on GCS which is
slower and not leveraging service accounts.
The underlying library Fog Google does support
the ability to use Google Storage JSON API
(https://github.com/fog/fog-google#storage)

This change adds support for the setting of the
relevant config options and passes these config
options to Fog Google so that we can use service
accounts.