Fix get_non_fingerprinted with dashed filenames
upload_files uses get_non_fingerprinted to determine if any
non-fingerprinted files need to be uploaded because there's a new
fingerprinted files. It does this by determining that if
assets/image-digest.png needs to be uploaded then assets/image.png
should also be uploaded, if it exists. But it faultily would determine
that if assets/great-image-digest.png needed to be uploaded then
assets/great.png should be uploaded if it exists instead of
assets/great-image.png