Drop trailing slash of prefix in #get_local_files
The bucket setup in `AssetSync::Storage#bucket` uses fuzzy matching when
`config.prefix` is present. This can present a problem in some cases, as
it doesn't allow for distinguishing between (e.g.) a bucket folder
called `assets/` and another folder called `assets-temp/`. A situation
could arise where the latter folder has thousands/millions of files and
everything slows to a crawl while we wait for the bucket object to
initialize.

This change allows developers to be more specific in their
`config.prefix` by using a trailing slash for their folder name.